refactor: DoingPage images move to assets
